+++ src/infra/outbound/deliver.ts
@@ -59,11 +59,13 @@ import { formatErrorMessage } from "../errors.js";
 import { throwIfAborted } from "./abort.js";
 import { resolveOutboundChannelMessageAdapter } from "./channel-resolution.js";
 import { resolveDeferredDeliveryAdmission } from "./deferred-delivery-admission.js";
-import {
-  applyFinalOutboundDeliveryPolicy,
-  applyOutboundDeliveryPolicy,
-  type OutboundDeliveryPolicyParams,
-} from "./deliver-policy.js";
+import type {
+  DeliverOutboundPayloadsCoreParams,
+  DeliverOutboundPayloadsParams,
+  OutboundDeliveryQueuePolicy,
+  PlatformSendRoute,
+} from "./deliver-contract.js";
+import { applyFinalOutboundDeliveryPolicy, applyOutboundDeliveryPolicy } from "./deliver-policy.js";
 import {
   OutboundDeliveryError,
   sessionKeyForDeliveryDiagnostics,
@@ -82,7 +84,6 @@ import {
   completeDurableDelivery,
   rejectDurableDelivery,
   suppressDurableDelivery,
-  type DurableDeliveryCompletion,
 } from "./delivery-completion.js";
 import { recordOutboundPolicySuppression } from "./delivery-policy-audit.js";
 import { releaseSpoolArtifacts, stageQueuePayloadMedia } from "./delivery-queue-media-spool.js";
@@ -98,7 +99,6 @@ import {
   markDeliveryPlatformSendDispatched,
   markDeliveryPlatformSendAttemptStarted,
   type QueuedReplyPayloadSendingHook,
-  type QueuedRenderedMessageBatchPlan,
   withActiveDeliveryClaim,
 } from "./delivery-queue.js";
 import type { OutboundDeliveryFormattingOptions } from "./formatting.js";
@@ -109,7 +109,6 @@ import {
   planOutboundTextMessageUnits,
   type OutboundMessageSendOverrides,
 } from "./message-plan.js";
-import type { DeliveryMirror } from "./mirror.js";
 import {
   completedOutboundAuditTerminals,
   emitOutboundAuditTerminals,
@@ -125,22 +124,16 @@ import {
 import { stripInternalRuntimeScaffolding } from "./protocol-scaffolding.js";
 import { createReplyToDeliveryPolicy } from "./reply-policy.js";
 import type { OutboundSendDeps } from "./send-deps.js";
-import type { OutboundSessionContext } from "./session-context.js";
 import type { OutboundChannel } from "./targets.js";
 
 export type { OutboundDeliveryResult } from "./deliver-types.js";
 export type { NormalizedOutboundPayload } from "./payloads.js";
 export type { OutboundSendDeps } from "./send-deps.js";
-
-export type OutboundDeliveryQueuePolicy = "required" | "best_effort";
-
-export type OutboundDeliveryIntent = {
-  id: string;
-  channel: Exclude<OutboundChannel, "none">;
-  to: string;
-  accountId?: string;
-  queuePolicy: OutboundDeliveryQueuePolicy;
-};
+export type {
+  DeliverOutboundPayloadsParams,
+  OutboundDeliveryIntent,
+  OutboundDeliveryQueuePolicy,
+} from "./deliver-contract.js";
 
 export type DurableFinalDeliveryRequirement = keyof NonNullable<
   ChannelDeliveryCapabilities["durableFinal"]
@@ -217,11 +210,6 @@ type ChannelHandler = {
     mediaUrl: string,
     overrides?: OutboundMessageSendOverrides,
   ) => Promise<OutboundDeliveryResult>;
-};
-
-type PlatformSendRoute = {
-  replyToId?: string | null;
-  threadId?: string | number | null;
 };
 
 type ChannelHandlerParams = {
@@ -752,76 +740,6 @@ async function persistQueuedPostSendState(params: {
     }
   }
 }
-
-type DeliverOutboundPayloadsCoreParams = OutboundDeliveryPolicyParams & {
-  cfg: OpenClawConfig;
-  channel: Exclude<OutboundChannel, "none">;
-  to: string;
-  accountId?: string;
-  payloads: ReplyPayload[];
-  replyToId?: string | null;
-  replyToMode?: ReplyToMode;
-  formatting?: OutboundDeliveryFormattingOptions;
-  threadId?: string | number | null;
-  identity?: OutboundIdentity;
-  deps?: OutboundSendDeps;
-  mediaAccess?: OutboundMediaAccess;
-  gifPlayback?: boolean;
-  forceDocument?: boolean;
-  replyPayloadSendingHook?: QueuedReplyPayloadSendingHook;
-  abortSignal?: AbortSignal;
-  bestEffort?: boolean;
-  onError?: (err: unknown, payload: NormalizedOutboundPayload) => void;
-  onPayload?: (payload: NormalizedOutboundPayload) => void;
-  /** @internal Reports the effective payload only after an identified platform send. */
-  onDeliveredPayload?: (payload: NormalizedOutboundPayload) => void;
-  onPayloadDeliveryOutcome?: (outcome: OutboundPayloadDeliveryOutcome) => void;
-  /** @internal Runs after each identified platform result, before further fallible work. */
-  onDeliveryResult?: (result: OutboundDeliveryResult) => Promise<void> | void;
-  /** @internal Persists ambiguous-send state immediately before platform I/O. */
-  onPlatformSendStart?: (route: PlatformSendRoute) => Promise<void>;
-  /** @internal Opaque durable intent id forwarded to provider reconciliation hooks. */
-  deliveryQueueId?: string;
-  /** @internal Stable producer id used to make queue creation idempotent across crashes. */
-  deliveryIntentId?: string;
-  /** @internal Serializable owner state finalized after live or recovered delivery. */
-  deliveryCompletion?: DurableDeliveryCompletion;
-  /** @internal Channel-valid id reserved before a correlated conversation turn is sent. */
-  preparedMessageId?: string;
-  /** @internal Recheck the concrete post-hook send shape before platform I/O. */
-  requiredUnknownSendReconciliation?: boolean;
-  /** @internal Caller preflight explicitly required provider unknown-send reconciliation. */
-  requireUnknownSendReconciliation?: boolean;
-  /** @internal Refresh durable timing before recipient-visible or finalizing platform I/O. */
-  onPlatformSendDispatch?: () => Promise<void>;
-  /** Session/agent context used for hooks and media local-root scoping. */
-  session?: OutboundSessionContext;
-  mirror?: DeliveryMirror;
-  silent?: boolean;
-  gatewayClientScopes?: readonly string[];
-  conversationReadOrigin?: "delegated" | "direct-operator";
-};
-
-/**
- * @deprecated Direct outbound delivery is compatibility/runtime substrate.
- * New message lifecycle code should use `sendDurableMessageBatch` from
- * `src/channels/message/send.ts` or `deliverInboundReplyWithMessageSendContext`
- * from `src/channels/turn/durable-delivery.ts`. Keep direct use only for
- * outbound substrate, recovery, and compatibility paths.
- */
-export type DeliverOutboundPayloadsParams = DeliverOutboundPayloadsCoreParams & {
-  /** @internal Skip write-ahead queue (used by crash-recovery to avoid re-enqueueing). */
-  skipQueue?: boolean;
-  /** @internal Recovery already ran provider admission after its pending-row re-read. */
-  deferredDeliveryAdmissionPassed?: true;
-  /** @internal State directory that owns the existing recovery queue entry. */
-  deliveryQueueStateDir?: string;
-  /** @internal Let recovery run commit hooks after it has acked the recovered queue entry. */
-  deferCommitHooks?: boolean;
-  queuePolicy?: OutboundDeliveryQueuePolicy;
-  renderedBatchPlan?: QueuedRenderedMessageBatchPlan;
-  onDeliveryIntent?: (intent: OutboundDeliveryIntent) => void;
-};
 
 type MessageSentEvent = {
   success: boolean;
