+++ src/infra/outbound/delivery-queue.ts
@@ -10,10 +10,7 @@ export {
   markDeliveryPlatformSendDispatched,
   markDeliveryPlatformSendAttemptStarted,
 } from "./delivery-queue-storage.js";
-export type {
-  QueuedReplyPayloadSendingHook,
-  QueuedRenderedMessageBatchPlan,
-} from "./delivery-queue-storage.js";
+export type { QueuedReplyPayloadSendingHook } from "./delivery-queue-storage.js";
 export {
   drainPendingDeliveries,
   recoverPendingDeliveries,
